Start writing an end-to-end test

+++ lib/streamlit/web/server/app_static_file_handler.py
@@ -28,7 +28,7 @@ _LOGGER = get_logger(__name__)
 # We agreed on these limitations for the initial release of static file sharing,
 # based on security concerns from the SiS and Community Cloud teams
 # The maximum possible size of single serving static file.
-MAX_APP_STATIC_FILE_SIZE = (config.get_option("server.maxStaticFileSize") or 0) * 1024 * 1024
+MAX_APP_STATIC_FILE_SIZE = int((config.get_option("server.maxStaticFileSize") or 0) * 1024 * 1024)
 
 SAFE_APP_STATIC_FILE_EXTENSIONS = tuple(config.get_option("server.allowedStaticFileExtensions"))
 if not SAFE_APP_STATIC_FILE_EXTENSIONS:  # if undefined or an empty list, use the default
@@ -58,7 +58,7 @@ class AppStaticFileHandler(tornado.web.StaticFileHandler):
 
         if (
             os.path.exists(full_path) 
-            and MAX_APP_STATIC_FILE_SIZE  # if 0 or None, there's no limit
+            and MAX_APP_STATIC_FILE_SIZE  # if 0, there's no limit
             and os.path.getsize(full_path) > MAX_APP_STATIC_FILE_SIZE
         ):
             raise tornado.web.HTTPError(
